refactor repo, controller merchant
