+++ src/main/java/com/example/trua_nay_an_gi/controller/merchant/MerchantController.java
@@ -32,4 +32,42 @@ public class MerchantController {
         merchant.setAccount(merchantOptional.get().getAccount());
         return new ResponseEntity<>(merchantService.save(merchant),HttpStatus.OK);
     }
+
+
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Merchant> findById(@PathVariable Long id) {
+        Optional<Merchant> merchantOptional = merchantService.findById(id);
+        if (!merchantOptional.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(merchantOptional.get(), HttpStatus.OK);
+    }
+
+//    @PutMapping("/{id}")
+//    public ResponseEntity<Merchant> updateMerchant(@PathVariable Long id, @RequestBody Merchant newMerchant) {
+//        Optional<Merchant> merchantOptional = merchantService.findById(id);
+//        if (!merchantOptional.isPresent()) {
+//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+//        }
+//        newMerchant.setId(id);
+//        return new ResponseEntity<>(merchantService.save(newMerchant), HttpStatus.OK);
+//    }
+//
+//    @PutMapping("/editMerchant/{id}")
+//    public ResponseEntity<Merchant> updateInformationMerchant(@PathVariable Long id, @RequestBody Merchant merchant) {
+//        Optional<Merchant> merchantOptional = merchantService.findById(id);
+//        if (!merchantOptional.isPresent()) {
+//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+//        }
+//        Merchant newMerchant = merchantOptional.get();
+//        newMerchant.setId(id);
+//        newMerchant.setName(merchant.getName());
+////        newMerchant.setDescription(merchant.getDescription());
+//        newMerchant.setAddress(merchant.getAddress());
+//        newMerchant.setPhone(merchant.getPhone());
+//        newMerchant.setOpenTime(merchant.getOpenTime());
+//        newMerchant.setCloseTime(merchant.getCloseTime());
+//        return new ResponseEntity<>(merchantService.save(newMerchant), HttpStatus.OK);
+//    }
 }
