+++ src/main/java/com/example/trua_nay_an_gi/service/merchant/MerchantService.java
@@ -1,7 +1,7 @@
 package com.example.trua_nay_an_gi.service.merchant;
 
 import com.example.trua_nay_an_gi.model.app_users.Merchant;
-import com.example.trua_nay_an_gi.repository.IMerchantRepository;
+import com.example.trua_nay_an_gi.repository.merchant.IMerchantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
